Update per CR review.
