+++ assets/js/modules/pagespeed-insights/components/dashboard/DashboardPageSpeedWidget/MetricSection.tsx
@@ -1,0 +1,65 @@
+/**
+ * DashboardPageSpeedWidget PDF MetricSection component.
+ *
+ * Site Kit by Google, Copyright 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies
+ */
+import { Text, View } from '@react-pdf/renderer';
+import type { ReactNode } from 'react';
+
+/**
+ * WordPress dependencies
+ */
+import { Fragment } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
+
+/**
+ * Internal dependencies
+ */
+import { styles } from './pdfStyles';
+
+interface MetricSectionProps {
+	title: string;
+	children: ReactNode;
+}
+
+export default function MetricSection( {
+	title,
+	children,
+}: MetricSectionProps ) {
+	return (
+		<Fragment>
+			<View style={ styles.headerRow }>
+				<View style={ styles.headerTitleCell }>
+					<Text style={ styles.headerSectionTitle }>{ title }</Text>
+				</View>
+				<View style={ styles.headerStrategyCell }>
+					<Text style={ styles.headerStrategyLabel }>
+						{ __( 'Mobile', 'google-site-kit' ) }
+					</Text>
+				</View>
+				<View style={ styles.headerStrategyCell }>
+					<Text style={ styles.headerStrategyLabel }>
+						{ __( 'Desktop', 'google-site-kit' ) }
+					</Text>
+				</View>
+			</View>
+			{ children }
+		</Fragment>
+	);
+}
